+++ lua/pomo/commands/util.lua
@@ -4,33 +4,41 @@ local pomo = require "pomo"
 local M = {}
 
 ---@param arg string
----@return pomo.Timer|?
-M.get_timer_from_arg = function(arg)
-  ---@type pomo.Timer|?
-  local timer
+---@return pomo.Timer[]|?
+M.get_timers_from_arg = function(arg)
+  ---@type pomo.Timer[]
+  local timers = {}
+
   if string.len(arg) > 0 then
+    -- Parse the argument to a timer ID.
     local timer_id = tonumber(arg)
     if timer_id == nil then
       log.error("invalid timer ID: '%s'", arg)
       return
     end
 
-    timer = pomo.get_timer(timer_id)
-    if not timer then
-      log.error("timer #%d is not active", timer_id)
-      return
+    if timer_id < 0 then
+      timers = pomo.get_all_timers()
     else
-      return timer
+      local timer = pomo.get_timer(timer_id)
+      if not timer then
+        log.error("timer #%d is not active", timer_id)
+        return
+      else
+        timers[#timers + 1] = timer
+      end
     end
   else
-    timer = pomo.get_latest()
+    local timer = pomo.get_latest()
     if not timer then
       log.error "there are no active timers"
       return
     else
-      return timer
+      timers[#timers + 1] = timer
     end
   end
+
+  return timers
 end
 
 return M